moved create report into own file, fixed tests
