+++ models/order.rb
@@ -1,5 +1,5 @@
-require_relative 'register'
-require_relative 'checkin'
+require_relative "register"
+require_relative "checkin"
 
 class Order
   attr_reader :name, :incomplete_order
@@ -36,37 +36,17 @@ class Order
     @items_array.each do |item|
       average += item.values[0].to_f
     end
-    return average / @items_array.count
+    average / @items_array.count
   end
 
-  def add_items_and_price(item, amount) 
+  def add_items_and_price(item, amount)
     price_total = (Register.find(item).price * amount.to_f)
+    price_total_convert = sprintf('%.2f', price_total)
     unless Order.find(@name).nil?
       @items_array += Order.find(@name).items_array
       Order.delete(Order.find(@name))
     end
-    @items_array << { item => "#{'%.2f' % price_total}" }
-  end
 
-  def self.create_report
-    order_text = []
-    @@all.each do |order|
-      if order.incomplete_order.nil?
-        items_string = ""
-        order.items_array.each do |item|
-          item.each do |key, value|
-            item_comma = ""
-            if order.items_array.count > 1 && (order.items_array.count)-1 > order.items_array.index(item)
-              item_comma = ","
-            end
-            items_string += " #{key} - $#{value}#{item_comma}"
-          end
-        end
-        order_text << "#{order.name}:#{items_string} | Average Order Value: $#{order.average_price.to_s}"
-      else
-        order_text << "#{order.name}: #{order.incomplete_order}"
-      end
-    end
-    order_text
+    @items_array << { item => price_total_convert }
   end
 end
