+++ spec/spec_helper.rb
@@ -3,3 +3,4 @@ require "rspec"
 require_relative "../models/register"
 require_relative "../models/checkin"
 require_relative "../models/order"
+require_relative "../create_report_helper"
